requirements.txt
python-telegram-bot[all]==20.0
requests==2.31.0
zxcvbn==4.4.28
pytz==2024.1
APScheduler==3.9.1
python-dotenv==1.0.0
openpyxl==3.1.0
